rename _utils.py -> utils.py because some functions are public
