+++ pdsparser/_PDS3_GRAMMAR.py
@@ -13,7 +13,7 @@ try:
 except ImportError:
     __version__ = 'Version unspecified'
 
-from ._utils import _based_int, _format_float, _is_identifier, _unwrap
+from .utils import _based_int, _format_float, _is_identifier, _unwrap
 
 ParserElement.set_default_whitespace_chars('')
 _WHITE = Suppress(Optional(Word(' \t')))
